Change logic for applying json field names in queries
The new logic allows query arguments to omit the name of the json field
only for the default (first) field name. This logic is simpler than the
original, and should allow queries to be more tolerant of changes to
joined-in schemas.

Multi-field constructors that include schemas now require a LnkedHashMap
rather than any generic Map, to ensure that insertion order is
maintained and the default field can be reliably identified.